+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Updated [windows-sys](https://github.com/microsoft/windows-rs/tree/master/crates/libs/sys) to
   0.34.0.
 - Moved `nstd_sys` contents into `nstd_env`, removing `nstd_sys` as a result.
+### `nstd.core`
+- Removed `nstd_core_abort` & `nstd_core_assert`.
 ### `nstd.events`
 - Cleaned `device_id` & `window_id`.
 - Made `NSTDDeviceID` `NSTDAny`.